chore: Refactor provider layer (#507)
* refactor: Review provider layer
